# contracts/labby-openapi.yaml
openapi: 3.1.0
info:
  title: Labby API
  version: 0.1.0
paths:
...
  /api/v2/labby/analytics/dashboard:
    get:
      tags:
      - sales-analytics
      summary: Dashboard
      operationId: dashboard_api_v2_labby_analytics_dashboard_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesDashboardStats'
      security:
      - HTTPBearer: []
  /api/v2/labby/sales/analytics/dashboard:
    get:
      tags:
      - sales-analytics
      summary: Dashboard
      operationId: dashboard_api_v2_labby_sales_analytics_dashboard_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesDashboardStats'
      security:
      - HTTPBearer: []
  /api/v2/labby/analytics/messages:
    get:
      tags:
      - sales-analytics
      summary: Message Volume
      operationId: message_volume_api_v2_labby_analytics_messages_get
      security:
      - HTTPBearer: []
      parameters:
      - name: period
        in: query
        required: false
        schema:
          type: string
          pattern: ^(7d|30d|90d)$
          default: 7d
          title: Period
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesMessageVolumeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/labby/sales/analytics/messages:
    get:
      tags:
      - sales-analytics
      summary: Message Volume
      operationId: message_volume_api_v2_labby_sales_analytics_messages_get
      security:
      - HTTPBearer: []
      parameters:
      - name: period
        in: query
        required: false
        schema:
          type: string
          pattern: ^(7d|30d|90d)$
          default: 7d
          title: Period
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SalesMessageVolumeResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/labby/analytics/activity:
    get:
      tags:
      - sales-analytics
      summary: Recent Activity
      operationId: recent_activity_api_v2_labby_analytics_activity_get
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 30
          minimum: 1
          default: 10
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SalesRecentActivityItem'
                title: Response Recent Activity Api V2 Labby Analytics Activity Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/labby/sales/analytics/activity:
    get:
      tags:
      - sales-analytics
      summary: Recent Activity
      operationId: recent_activity_api_v2_labby_sales_analytics_activity_get
      security:
      - HTTPBearer: []
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 30
          minimum: 1
          default: 10
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SalesRecentActivityItem'
                title: Response Recent Activity Api V2 Labby Sales Analytics Activity
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SalesDashboardStats:
      properties:
        mensagens_hoje:
          type: integer
          title: Mensagens Hoje
        mensagens_semana:
          type: integer
          title: Mensagens Semana
        contatos_total:
          type: integer
          title: Contatos Total
        conversas_abertas:
          type: integer
          title: Conversas Abertas
        campanhas_ativas:
          type: integer
          title: Campanhas Ativas
        taxa_resposta:
          type: number
          title: Taxa Resposta
      type: object
      required:
      - mensagens_hoje
      - mensagens_semana
      - contatos_total
      - conversas_abertas
      - campanhas_ativas
      - taxa_resposta
      title: SalesDashboardStats
...
    SalesMessageVolumeItem:
      properties:
        date:
          type: string
          title: Date
        enviadas:
          type: integer
          title: Enviadas
        recebidas:
          type: integer
          title: Recebidas
      type: object
      required:
      - date
      - enviadas
      - recebidas
      title: SalesMessageVolumeItem
    SalesMessageVolumeResponse:
      properties:
        period:
          type: string
          enum:
          - 7d
          - 30d
          - 90d
          title: Period
        data:
          items:
            $ref: '#/components/schemas/SalesMessageVolumeItem'
          type: array
          title: Data
      type: object
      required:
      - period
      - data
      title: SalesMessageVolumeResponse
...
    SalesRecentActivityItem:
      properties:
        tipo:
          type: string
          enum:
          - conversa
          - campanha
          title: Tipo
        titulo:
          type: string
          title: Titulo
        descricao:
          anyOf:
          - type: string
          - type: 'null'
          title: Descricao
        canal:
          anyOf:
          - type: string
          - type: 'null'
          title: Canal
        timestamp:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Timestamp
        link_id:
          type: string
          format: uuid
          title: Link Id
        status:
          type: string
          title: Status
        aguardando_humano:
          type: boolean
          title: Aguardando Humano
      type: object
      required:
      - tipo
      - titulo
      - link_id
      - status
      - aguardando_humano
      title: SalesRecentActivityItem
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer